Deleted a bug when library crashed while multiplications in GF(2^n) runs.
The code was cleaned and tested under MINGW32 and MSVC on 32bit
Intel architecture

+++ examples/example-oid.c
@@ -28,7 +28,7 @@
   }
 
  /* выводим информацию об общем количестве OID библиотеки */
- printf("total count of oid's: %ld\n", ak_libakrypt_oids_count( ));
+ printf("total count of oid's: %lu\n", (unsigned long int) ak_libakrypt_oids_count( ));
 
  /* ищем OID по его имени */
   printf("\nsearch results:\n");
